langextract-vllm/requirements.txt
langextract
vllm
pyyaml
